Fix compilation with g++-4.6.
git-svn-id: https://jsonrpc-cpp.svn.sourceforge.net/svnroot/jsonrpc-cpp/trunk@58 bbb04abe-9aa6-4717-80b1-ed36530a708e

+++ src/jsonrpc_handler.cpp
@@ -199,8 +199,8 @@ namespace Json
       if(root.isArray())
       {
         /* batched call */
-        size_t i = 0;
-        size_t j = 0;
+        unsigned int i = 0;
+        unsigned int j = 0;
         
         for(i = 0 ; i < root.size() ; i++)
         {
